requirements.txt
torch>=2.0.0
transformers>=4.30.0
ultralytics>=8.0.0
Pillow>=9.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
accelerate>=0.20.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
